+++ packages/@runlightyear/github/src/webhooks/repositoryWebhooks/createRepositoryWebhook.ts
@@ -1,4 +1,7 @@
-import { HttpProxyResponse } from "@runlightyear/lightyear";
+import {
+  HttpProxyResponse,
+  HttpProxyResponseError,
+} from "@runlightyear/lightyear";
 import { GitHub } from "../../GitHub";
 import WebhookEvent from "../../types/WebhookEvent";
 import WebhookConfig from "../../types/WebhookConfig";
@@ -73,22 +76,40 @@ const createRepositoryWebhook =
   ): Promise<CreateRepositoryWebhookResponse> => {
     const { owner, repo, name, config, events, active } = props;
 
-    return self.post({
-      url: `/repos/${owner}/${repo}/hooks`,
-      data: {
-        name,
-        config: config && {
-          url: config.url,
-          content_type: "json",
-          secret: config.secret,
-          insecure_ssl: config.insecureSSL,
-          token: config.token,
-          digest: config.digest,
+    try {
+      console.debug("about to call createRepositoryWebhook");
+      const response = await self.post({
+        url: `/repos/${owner}/${repo}/hooks`,
+        data: {
+          name,
+          config: config && {
+            url: config.url,
+            content_type: "json",
+            secret: config.secret,
+            insecure_ssl: config.insecureSSL,
+            token: config.token,
+            digest: config.digest,
+          },
+          events,
+          active,
         },
-        events,
-        active,
-      },
-    });
+      });
+
+      console.debug("got the response");
+
+      return response;
+    } catch (error) {
+      console.debug("Got an error");
+      if (error instanceof HttpProxyResponseError) {
+        if (error.response.status === 404) {
+          console.error(
+            "Check that the owner and repo are specified correctly. The owner parameter is the account owner of the repository (case insensitive). The repo parameter is name is the name of the repository without the .git extension (case insensitive)."
+          );
+        }
+      }
+
+      throw error;
+    }
   };
 
 export default createRepositoryWebhook;