+++ packages/@runlightyear/airtable/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/airtable
 
+## 0.3.3
+
+### Patch Changes
+
+- Updated dependencies [7469fbe82]
+  - @runlightyear/lightyear@0.33.0
+
 ## 0.3.2
 
 ### Patch Changes