+++ packages/@runlightyear/postmark/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/postmark
 
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies [7469fbe82]
+  - @runlightyear/lightyear@0.33.0
+
 ## 0.2.2
 
 ### Patch Changes